D2M: ApplyHostMemrefCallingConvention pass (#4071)
### Ticket
N/A

### Problem description
We need to annotate host memref types with strides in order to represent
its memory alignment. The assumption is that memrefs with 3 and more
dimensions have rows that are packed and therefore don't need tile
height alignment and columns are not packed and therefore need tile
width alignment.

For example, for a memref with logical shape of 128x2x2 the strides will
be [64, 32, 1] to match the memory representation of rows of 32 columns
with 2 values within each one.

### What's changed
Added new `ApplyHostMemrefCallingConvention` pass after
`ConvertTTIRToTTMetal` pass that applies `UniformTypeRewriter` to update
host `memref` types. Host `memref` type is distinguished by the absence
of memory space attribute. In addition to `UniformTypeRewriter` there is
also `MemrefGlobalRewriter` to handle rewriting `memref` type attribute
inside of `memref:global`.