Fixing bugs with lowering TTIR to LINALG and adding tests for them. (#4169)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/4168

### Problem description
TTIR.matmul had problems when lowering to linalg due to reshaping.
TTIR.add had problems when adding two tensors with different ranks as no
broadcast is used.

### What's changed
For ```ttir.matmul```: Different methods for reshaping lhs and rhs are
used. Method for lhs does not return errors, while the other one does.
Changed reshaping of rhs to be equivalent to the reshaping of lhs.
For ```ttir.add```: Added broadcasting for elementwise binary operations
if ranks of input are not the same.

### Checklist
- [x] New/Existing tests provide coverage for changes