Refactor `TTNNOpModel` (#4230)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/4229

### Problem description
`TTNNOpModel` contains a lot of code repetition, which makes onboarding
new ops more tedious, and makes future maintenance much harder.

### What's changed
- Changed approach from namespace-based to template struct one.
Namespace doesn't convey any info, beside convention. This approach
enables writing type-safe generic code, which is mostly visible in how
tests are written now.
- Removed `name` parameter from `operation::getOpConstraints` and
`operation::getOpRuntime`; this parameter wasn't used anywhere
- Removed `outputShape` parameter from all ops, `ReshapeOp` is an
exception because there `outputShape` is used to infer an attribute of
an op, it's not used for specifying `TensorSpec` anywhere.

This is not a final state of refactoring; there are still things that
need to be done, but since this is already a pretty big change, I
decided to take a break here for review.

TODO:
- adding eltwise and reduction ops is trivial now; I didn't want to
introduce that change in this PR, since this is refactoring, but plan to
do so in subsequent PR

### Checklist
- [x] New/Existing tests provide coverage for changes