Return empty data vector (#4238)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/4242

### Problem description
Runtime provides 'getTensorDataBuffer' function to extract the data from
ttnn::Tensor and return it as data vector. However, it assumes that the
underlying tensor will always contain some data and assert otherwise.
This assumption is not true in case of empty tensor.

### What's changed
Update the functionality to return empty data vector in case of empty
'ttnn::Tensor'