[Optimizer] Respect override when checking interleaved to sharded (#4188)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-mlir/issues/3749)

### Problem description
When checking whether a reshard must be inserted before the first op in
the L1 chain, or whether it supports interleaved input with sharded
output, we assumed that the op has tiled input. However, if the op
before it has row major output, that makes its input row major as well
and there is a possibility it does not support sharded output in that
case.

### What's changed
The set of ops with row major output overrides is created in
`TTNNOptimizer`, passed through all intermediate functions, and finally
used in `ShardSolver::preprocessFirstOp`.

### Checklist
- [ ] New/Existing tests provide coverage for changes