Update open mesh device API to optional mesh shape (#4432)
### Problem description
Mesh shape is now an optional parameter when creating a mesh device in
metal. When no mesh shape is passed in, it will default to the topology
mesh shape (1x2 on n300, 2x4 on t3k etc.).

### What's changed
Made mesh shape optional on our end as well, and moved it into
`MeshDeviceOptions`. Also updated to pass `std::nullopt` when creating
system descriptor so we get a descriptor describing the entire topology.
With this hopefully it makes it easier to implement topology APIs in
PJRT as well.

I will update the FEs accordingly once this lands.

### Checklist
- [X] New/Existing tests provide coverage for changes