Adding bfp8 support to tt-mlir (#4325)
Today we normalize all types in graph by converting mlir type to mlir
type which is supported by tt hardware. These supported types (i.e
bf16/f32/f16) are well supported by runtime stack and frontends. We lack
support for block types on runtime side since they are specific only to
tt hardware. In particular we lack some way to create bfp8 tensor and
memcpy bfp8 tensor (return as output).

To workaround this we will leave function inputs and outputs in same
data type as they came. In case of bfp8_b inputs and outputs will be
left in bf16 type and we will insert typecast at the start and end of
function body to convert bf16 into bfp8_b and back to bf16.

North star is to have all conversions on types behave like this so that
FE won't have to keep track of what was the original type (before the
compiler was invoked) and what is converted type.

Will add emitc test in next commit.